Update create.sql
Changes ids from Int to BigInt

+++ database/create/create.sql
@@ -3,13 +3,13 @@ DROP DATABASE IF EXISTS my_finances;
 CREATE DATABASE my_finances;
 
 CREATE TABLE partner (
-	id INT AUTO_INCREMENT PRIMARY KEY,
+	id BIGINT AUTO_INCREMENT PRIMARY KEY,
 	last_name VARCHAR(255) NOT NULL,
 	first_name VARCHAR(255)
 );
 
 CREATE TABLE property (
-	id INT AUTO_INCREMENT PRIMARY KEY,
+	id BIGINT AUTO_INCREMENT PRIMARY KEY,
 	name VARCHAR(255) NOT NULL,
 	description TEXT,
 	runtime INT,
@@ -20,16 +20,16 @@ CREATE TABLE property (
 );
 
 CREATE TABLE account (
-	id INT AUTO_INCREMENT PRIMARY KEY,
+	id BIGINT AUTO_INCREMENT PRIMARY KEY,
 	position INT,
-	master_id INT,
+	master_id BIGINT,
 	acc_name VARCHAR(255),
 	FOREIGN KEY (master_id) REFERENCES account(id)
 );
 
 CREATE TABLE transaction(
-	id INT AUTO_INCREMENT PRIMARY KEY,
-	account_id INT,
+	id BIGINT AUTO_INCREMENT PRIMARY KEY,
+	account_id BIGINT,
 	value_original DECIMAL(10, 2),
 	currency_value_original VARCHAR(3),
 	date DATE,
@@ -39,32 +39,32 @@ CREATE TABLE transaction(
 );
 
 CREATE TABLE sentencematurity (
-	id INT AUTO_INCREMENT PRIMARY KEY,
+	id BIGINT AUTO_INCREMENT PRIMARY KEY,
 	name VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE sentence (
-	id INT AUTO_INCREMENT PRIMARY KEY,
+	id BIGINT AUTO_INCREMENT PRIMARY KEY,
 	description VARCHAR(255),
 	sentencematurity_id INT,
 	FOREIGN KEY (sentencematurity_id) REFERENCES sentencematurity(id)
 );
 
 CREATE TABLE word (
-	id INT AUTO_INCREMENT PRIMARY KEY,
-	sentence_id INT,
-	account_id INT,
+	id BIGINT AUTO_INCREMENT PRIMARY KEY,
+	sentence_id BIGINT,
+	account_id BIGINT,
 	date DATE,
 	value DECIMAL(10, 2),
 	currency_value VARCHAR(3),
-	partner_id INT,
+	partner_id BIGINT,
 	FOREIGN KEY (sentence_id) REFERENCES sentence(id),
 	FOREIGN KEY (account_id) REFERENCES account(id),
 	FOREIGN KEY (partner_id) REFERENCES partner(id)
 );
 
 CREATE TABLE trf_transaction_sentence (
-	sentence_id INT,
-	transaction_id INT,
+	sentence_id BIGINT,
+	transaction_id BIGINT,
 	PRIMARY KEY (sentence_id, transaction_id)
 );
